requirements.txt
discord.py==2.3.2
python-decouple==3.8
pytz